Add files via upload
An application that enables registered/authorised users to log in and create a new customer account and BankAccount. 
This uses a four-digit account numbers, but the program does not have to enforce this rule.  
Containts a struct so that it contains public data fields that hold the int account number and double account balance and customer name and password.  
The main() function that declares a BankAccount and allows you to enter data values for the BankAccount fields.  
Contains a logging in part of the system, which will prompt users for their account numbers and name to enable them access to the program functionality such as viewing their bank balance.
**At this stage, password validation, or data persistency is not accurate because I have not given full attention to that (ie storing data for future usages). This will be developed next.