BankAccount_Stage_2
Prompt the user for account numbers and beginning balances for his/her BankAccounts. 
For data entry validation, the bank account number should be between 1000 and 9999; Issues an error message if the user does not enter a valid account number. The balance should not be negative; issues an error message if it is. Continue with the program only if the account numbers and balances are valid for both.  
Display the full account numbers and starting balances for the two accounts. 
Prompt the user for a dollar amount to be transferred from the first account to the second account. Issues an error message if the two accounts have the same account number, and will not carry out the transfer of funds.  
Issues an error message if the requested transfer causes the first account to fall below $0.00, and will not make the transfer.
Issues a warning message if the transfer causes the balance in the first account to drop below $10.00, but make the transfer. 
Issues a warning message if the transfer causes the balance in the second account to be greater than $100,000.00, which is the highest amount that is federally insured. 
Display the ending balances after the transfer amount has been deducted from the first account and added to the second. 

** For now, when an error message is issued, the program terminates gracefully. For warning messages, the program continues to execute.